Another Profile View Update
So it was crashing due to one line of code:
@interface ProfileViewController (CameraDelegateMethods)

Apparently the CameraDelegateMethods does not need to be there, at all.
I believe I inserted it their from the first tutorial I tried following
which wasn't complete, so I should have removed it.

For some reason it only caused crashes once the core data stuff was
implemented though… Either way, it's all working now - the core data
name and tagline saving, as well as the camera taking a photo and
storing it in the image view.